+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [2.8.0-dev.3](https://github.com/PocketMiner82/SongGuess/compare/2.8.0-dev.2...2.8.0-dev.3) (2026-01-24)
+
+
+### Bug Fixes
+
+* move ToastError after TopBar component ([309c8cd](https://github.com/PocketMiner82/SongGuess/commit/309c8cd453498a256abedf6153ed28a67daf4b24))
+
+# [2.8.0-dev.2](https://github.com/PocketMiner82/SongGuess/compare/2.8.0-dev.1...2.8.0-dev.2) (2026-01-24)
+
+
+### Bug Fixes
+
+* **BaseConfig:** better default values ([d775684](https://github.com/PocketMiner82/SongGuess/commit/d775684ad9d5c59691e6c7bdb1ee1da1cb73e213))
+
+# [2.8.0-dev.1](https://github.com/PocketMiner82/SongGuess/compare/2.7.0...2.8.0-dev.1) (2026-01-24)
+
+
+### Features
+
+* add better, unified error components ([d87d430](https://github.com/PocketMiner82/SongGuess/commit/d87d4302cc2a1969c75a40e04fde36f5d001d1ff))
+
 # [2.7.0](https://github.com/PocketMiner82/SongGuess/compare/2.6.0...2.7.0) (2026-01-23)